requirements.txt
# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# Ollama integration
ollama==0.1.6

# Vector store
chromadb==0.4.22
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.4
pdfplumber==0.10.3
python-docx==1.1.0
python-magic==0.4.27

# CLI framework
click==8.1.7
rich==13.7.0
prompt-toolkit==3.0.43

# API framework (optional, for API server mode)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tiktoken==0.5.2
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
